Name upstream and debian branches properly
Upstream keeps it's changes on "master". This may be the default for
`gbp`, but may *not* be. Explicitly set the upstream branch just to be
sure.

The Debian branch shouldn't be called "trunk" (or "main" or "master").
I'm calling it "deb" in case a real Debian maintainer picks up the
upstream project and wants to use the name "debian". It's configurable,
so hardly critical, but I may as well be a good neighbor and leave it
unused.