Adjust the copyright file a bit...
I'm going to mark Geoffrey as having the copyright range up to 2025
because that's when I'm actually assembling this branch. Not like this
needs to be made as-if it was done at that point.

The previous copyright file indicated "LGPL" was an option. I think I
had done that because of the "GPLv3 *or later*" phrasing. I don't think
that's how it works -- the LGPL is not a "newer version" of the GPL.